+++ modules/l7_load_balancers/examples/simple-lbaas-on-existing-vcn/l7_load_balancers.tf
@@ -54,6 +54,7 @@ locals {
       rule_sets                      = l7lb_value.rule_sets
       certificates                   = l7lb_value.certificates
       listeners                      = l7lb_value.listeners
+      security                       = l7lb_value.security
     }
   } : {} : {}